backend/Recon-Analyzer/requirements.txt
flask
requests
python-dotenv
beautifulsoup4
Flask-Cors